feat(GameLogic): Se agrega recibir daño y se arregla parar de disparar
Estaba mal en handler event disparar, no paraba nunca.
Por otro lado agregue el estado de estar siendo lastimado para los zombies.
Tendria que agregar un cooldown para la recarga del arma, la hace instantanea.
Ahora el shoot se ve limpio porque antes se le acaban las balas muy rapido y recargaba todo el tiempo. Ahora al durarle más (las cambie en el yaml) recarga menos veces.

+++ Server/include/GameLogic/Weapons/idfweapon.h
@@ -9,14 +9,14 @@ Debe recargar el arma cada 50 rafagas. */
 
 class IdfWeapon : public Weapon {
 private:
-    uint8_t ammo;
-    uint8_t actual_ammo;
+    uint16_t ammo;
+    uint16_t actual_ammo;
     double damage;
     double scope;
     double damage_reduction_coef;
 
 public:
-    explicit IdfWeapon(uint8_t ammo, double damage, double scope, double reduction);
+    explicit IdfWeapon(uint16_t ammo, double damage, double scope, double reduction);
 
     bool shoot(
     Position& from,
