+++ Server/include/GameLogic/Weapons/scoutweapon.h
@@ -10,14 +10,14 @@ Se recarga cada 20 balas. */
 
 class ScoutWeapon : public Weapon {
 private:
-    uint8_t ammo;
-    uint8_t actual_ammo;
+    uint16_t ammo;
+    uint16_t actual_ammo;
     double damage;
     double scope;
     double damage_reduction_coef;
 
 public:
-    explicit ScoutWeapon(uint8_t ammo, double damage, double scope, double reduction);
+    explicit ScoutWeapon(uint16_t ammo, double damage, double scope, double reduction);
 
     bool shoot(
     Position& from,
